Added a download button.
